overall code cleanup and resolve previous feedback
