+++ frontend/src/components/EventDetail/EventDetail.jsx
@@ -1,34 +1,12 @@
 import "./EventDetail.css"
-import { getReadableDateStr, getReadableTimeStr } from "../../utils/utils";
+import { handleEventDelete, getReadableDateStr, getReadableTimeStr } from "../../utils/utils";
 function EventDetail ({onClose, content, refetchEvents, handleEdit}) {
-
-    const handleEventDelete = async () => {
-        try{
-            const backendUrlAccess = import.meta.env.VITE_BACKEND_ADDRESS;
-            const options = {
-            method: 'DELETE',
-            headers: {
-                'Accept': 'application/json',
-                'Content-Type': 'application/json'},
-            credentials: 'include'
-              };
-            const response = await fetch(`${backendUrlAccess}/calendar/events/${content.id}`,options);
-            if (!response.ok) {
-              throw new Error('Something went wrong!');
-            }
-            refetchEvents();
-            onClose();
-          }
-          catch(error) {
-            console.error(error);
-          }
-    }
 
     return (
         <div className="modal-overlay">
             <div className="modal-content">
                     <button onClick={handleEdit}>Edit</button>
-                    <button onClick={() => handleEventDelete()}>Delete</button>
+                    <button onClick={() => handleEventDelete(content, onClose, refetchEvents)}>Delete</button>
                     <h1>{content.title}</h1>
 
                     <p>{getReadableDateStr(content.start)}</p>
