+++ frontend/src/components/SignUpForm/SignUpForm.jsx
@@ -3,6 +3,7 @@ import "./SignUpForm.css"
 import { UserContext } from '../../UserContext.js';
 import { useNavigate, Link} from 'react-router-dom';
 import {validate as validateEmail} from 'email-validator';
+import { handleSignUp } from "../../utils/utils.js";
 
 function SignUpForm (){
     const [email, setEmail] = useState("");
@@ -49,40 +50,7 @@ function SignUpForm (){
               password: password
           }
 
-          try {
-            // Make the signup API request
-            const response = await fetch(`http://localhost:3000/auth/registration`, {
-              method: 'POST',
-              headers: {
-                'Content-Type': 'application/json',
-              },
-              body: JSON.stringify(userObj),
-              credentials: 'include'
-            });
-
-            if (response.ok) {
-              const data = await response.json();
-              const loggedInUser = data.user;
-
-              // Reset form fields
-              setFirstName('');
-              setLastName('');
-              setEmail('');
-              setPassword('');
-
-              // Update the user context
-              updateUser(loggedInUser);
-
-              // Navigate to the home page after successful login
-              navigate('/');
-            } else {
-              // Handle signup failure case
-              alert('Signup failed');
-            }
-          } catch (error) {
-            // Handle any network or API request errors
-            alert('Signup failed: ' + error);
-          }
+          handleSignUp(userObj, updateUser, navigate, setFirstName, setLastName, setEmail, setPassword);
       } else {
         alert("Please fill out all fields");
       }
